relayer/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
web3>=6.15.0
python-dotenv>=1.0.0
pydantic>=2.6.0
httpx>=0.27.0
numpy>=1.26.0
apscheduler>=3.10.0
pytest>=8.0.0
